requirements.txt
sqlalchemy==1.3.20
python-dotenv==1.0.0
pymysql==1.1.0
black==22.8.0